# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant IAM API
  version: 1.0.0
  description: API para gestión de identidad y acceso multi-tenant
servers:
  - url: /iam/api/v1
    description: Base URL para la API v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticación
    tenantId:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: ID del tenant para operaciones multi-tenant
  parameters:
...
    PageParam:
      name: page
      in: query
      description: "Número de página (default: 1)"
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: "Tamaño de página (default: 10, max: 100)"
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortByParam:
      name: sort_by
      in: query
      description: "Campo de ordenamiento (default: created_at)"
      required: false
      schema:
        type: string
        default: "created_at"
    SortDirParam:
      name: sort_dir
      in: query
      description: "Dirección de ordenamiento (asc o desc, default: desc)"
      required: false
      schema:
        type: string
        enum: ["asc", "desc"]
        default: "desc"

    # Parámetros específicos para usuarios
    UserTenantIdParam:
      name: tenant_id
      in: query
      description: ID del tenant
      required: false
      schema:
        type: string
        format: uuid
    UserStatusParam:
      name: status
      in: query
      description: Estado del usuario
      required: false
      schema:
        type: string
        enum: ["ACTIVE", "INACTIVE", "SUSPENDED", "DELETED"]
    UserRoleIdParam:
      name: role_id
      in: query
      description: ID del rol
      required: false
      schema:
        type: string
        format: uuid
    UserEmailParam:
      name: email
      in: query
      description: Búsqueda por email (LIKE)
      required: false
      schema:
        type: string
    UserFirstNameParam:
      name: first_name
      in: query
      description: Búsqueda por nombre (LIKE)
      required: false
      schema:
        type: string
    UserLastNameParam:
      name: last_name
      in: query
      description: Búsqueda por apellido (LIKE)
      required: false
      schema:
        type: string
    UserProviderParam:
      name: provider
      in: query
      description: Proveedor de autenticación
      required: false
      schema:
        type: string
        enum: ["LOCAL", "GOOGLE"]

    # Parámetros específicos para roles
...
  schemas:
...
    ListResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total_count:
          type: integer
          description: Total de elementos
        page:
          type: integer
          description: Página actual
        page_size:
          type: integer
          description: Tamaño de página
        total_pages:
          type: integer
          description: Total de páginas

    # Respuestas específicas usando ListResponse
    UserListResponse:
      allOf:
        - $ref: '#/components/schemas/ListResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/User'
...
    User:
      type: object
      required:
        - email
        - first_name
        - last_name
        - tenant_id
        - role_id
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        tenant_id:
          type: string
          format: uuid
        role_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED, DELETED]
        provider:
          type: string
          enum: [LOCAL, GOOGLE]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /users:
    get:
      summary: Listar usuarios con filtros y paginación
      tags: [Users]
      security:
        - BearerAuth: []
        - tenantId: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortDirParam'
        - $ref: '#/components/parameters/UserTenantIdParam'
        - $ref: '#/components/parameters/UserStatusParam'
        - $ref: '#/components/parameters/UserRoleIdParam'
        - $ref: '#/components/parameters/UserEmailParam'
        - $ref: '#/components/parameters/UserFirstNameParam'
        - $ref: '#/components/parameters/UserLastNameParam'
        - $ref: '#/components/parameters/UserProviderParam'
      responses:
        '200':
          description: Lista paginada de usuarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          description: Parámetros de filtrado inválidos
        '401':
          description: No autorizado
        '403':
          description: Sin permisos para acceder a usuarios

    post:
      summary: Crear nuevo usuario
      tags: [Users]
      security:
        - BearerAuth: []
        - tenantId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /users/{id}:
    get:
      summary: Obtener usuario por ID
      tags: [Users]
      security:
        - BearerAuth: []
        - tenantId: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    delete:
      summary: Eliminar usuario
      tags: [Users]
      security:
        - BearerAuth: []
        - tenantId: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Usuario eliminado
...